refactor: adjust ast related files
