+++ src/view/components/StyleInformation.vue
@@ -18,7 +18,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import { RuleForPrint } from '@/parser/style'
+import { RuleForPrint } from '@/parser/style/types'
 
 export default Vue.extend({
   name: 'StyleInformation',
