+++ types/babylon.d.ts
@@ -1,0 +1,3 @@
+import * as t from '@babel/types'
+
+export function parse(code: string, options?: any): t.File
